Move DateTooltip to it's on file
